스케줄 상세화면 기능 - (하트) 좋아요 기능 구현

+++ src/main/java/com/example/recard/dto/ScheduleDto.java
@@ -18,7 +18,7 @@ public class ScheduleDto {
     private Long schedule_id;
     private String title;
     private String content;
-    private Long likeCount;
+    private Integer likeCount;
 
     private int state;
